+++ network/manager_message.go
@@ -1,5 +1,8 @@
 package network
 
+// ManagerMessage is a message sent meant for parsing by the ConnectionManager
+// It is identical to ConsoleMessage - This is so that either side can parse messages
+// without having to know the other side.
 type ManagerMessage struct {
 	Type        string `json:"type"`
 	Message     string `json:"message"`
@@ -13,6 +16,7 @@ type ManagerMessage struct {
 type MessageType string
 
 // MessageTypes are the types of messages
+
 const (
 	MessageTypeChat MessageType = "chat"
 )
