+++ ui/types/console_messages.go
@@ -2,6 +2,8 @@ package types
 
 import "encoding/json"
 
+// ConsoleMessage is a message that is meant for parsing by the Console
+// They can be sent by the ConnectionManager or by windows into the Console
 type ConsoleMessage struct {
 	Type        string `json:"type"`
 	Message     string `json:"message"`
@@ -11,6 +13,7 @@ type ConsoleMessage struct {
 	WindowID    int    `json:"window_id"`
 }
 
+// String returns a JSON string representation of the ConsoleMessage
 func (c *ConsoleMessage) String() string {
 	out, _ := json.Marshal(c)
 	return string(out)
